vercel-solver/requirements.txt
quart==0.20.0
hypercorn==0.17.3
python-dotenv==1.0.1
requests==2.32.3
pytest==8.3.5
